(PC-22795)[PRO] refactor: add full-validate and remove other full validate like icon
