+++ pro/src/screens/OfferIndividualConfirmation/OfferIndividualConfirmation.module.scss
@@ -1,5 +1,6 @@
 @use "styles/mixins/_rem.scss" as rem;
 @use "styles/mixins/_fonts.scss" as fonts;
+@use "styles/variables/_colors.scss" as colors;
 
 .confirmation-container {
   text-align: center;
@@ -7,6 +8,7 @@
 
 .validate-icon {
   height: rem.torem(56px);
+  color: colors.$green-valid;
 }
 
 .pending-icon {
